requirements.txt
Django>=5.2.4
python-telegram-bot>=20.0
python-dotenv>=1.0.0
asgiref>=3.7.0
requests>=2.31.0
qrcode[pil]>=7.4.0 